[ADD] implement usePermissions hook for role-based access control in available resources

+++ src/autoCall/components/Table/columnsDisponibles.tsx
@@ -2,13 +2,14 @@
 
 import { ColumnDef } from "@tanstack/react-table"
 import { format } from "date-fns";
-import { newDataDisponible } from "@/autoCall/components/Interfaces/Disponible"
+import { NewDataDisponible } from "@/autoCall/components/Interfaces/Disponible"
 import { Button } from "@/components/ui/button"
 import { SorterIcon } from "./utils"
 import { Dialog, DialogClose, DialogContent, DialogDescription, DialogFooter, DialogHeader, DialogTitle, DialogTrigger } from "@/components/ui/dialog";
 import { UserX } from "lucide-react";
 import { deleteData } from "@/api/apiCrud";
 import { toast } from "sonner"
+import { usePermissions } from "@/auth/Permissions/usePermissions";
 
 function formatCell(cellPhone:number){
     const formatted = cellPhone
@@ -23,9 +24,10 @@ function formatCell(cellPhone:number){
 
 
 export const columnsDisponibles = (fetchData: () => Promise<void>) => {
+  const { hasPermission } = usePermissions()
 
 
-  const onDelete = async (data:newDataDisponible)=>{  
+  const onDelete = async (data:NewDataDisponible)=>{  
     const promiseDeleteUser = new Promise(async (resolve, reject)=>{
       try {
         const response = await deleteData('disponibles', data.Torre, data.RowKey)        
@@ -107,35 +109,38 @@ export const columnsDisponibles = (fetchData: () => Promise<void>) => {
       return <div>{format(date as Date, "dd/MM/yyy")}</div>
     }
   },
+
   {
     id: "actions",
     header: "Acciones",    
     cell: ({ row }) => {
       const disponible = row.original;
-      return (
-        <Dialog>
-        <DialogTrigger><UserX/></DialogTrigger>
-        <DialogContent>
-          <DialogHeader>
-            <DialogTitle>¿Estás completamente seguro?</DialogTitle>
-            <DialogDescription>
-              Esta acción no se puede deshacer. Eliminará permanentemente su usuario.
-            </DialogDescription>
-          </DialogHeader>
-        <DialogFooter>          
-          <DialogClose asChild>
-            <Button type="button" variant="secondary">
-              Cancelar
-            </Button>
-          </DialogClose>
-          <DialogClose asChild>
-            <Button type="button" onClick={()=>onDelete(disponible)}>Aceptar</Button>
-          </DialogClose>
-        </DialogFooter>
-        </DialogContent>
-      </Dialog>      
-      );
+      if(hasPermission('availables', 'delete')){
+        return (
+          <Dialog>
+          <DialogTrigger><UserX/></DialogTrigger>
+          <DialogContent>
+            <DialogHeader>
+              <DialogTitle>¿Estás completamente seguro?</DialogTitle>
+              <DialogDescription>
+                Esta acción no se puede deshacer. Eliminará permanentemente su usuario.
+              </DialogDescription>
+            </DialogHeader>
+          <DialogFooter>          
+            <DialogClose asChild>
+              <Button type="button" variant="secondary">
+                Cancelar
+              </Button>
+            </DialogClose>
+            <DialogClose asChild>
+              <Button type="button" onClick={()=>onDelete(disponible)}>Aceptar</Button>
+            </DialogClose>
+          </DialogFooter>
+          </DialogContent>
+        </Dialog>      
+        );
+      }
     },
   },
-] as ColumnDef<newDataDisponible>[]
+] as ColumnDef<NewDataDisponible>[]
 }
